move cache handling to seperate file
